(Feature) Move r3_dnn_apply from scan_batteries to lib

+++ lib/evaluate_models.py
@@ -54,4 +54,12 @@ if __name__ == '__main__':
         Popen(commands[1], shell=True).wait()
         print('\n\nevaluate_models.py: processing in vivo for model', model_index + 1, 'of', num_models, ':', os.path.basename(model_folder), '\n\n')
         Popen(commands[2], shell=True).wait()
-        shutil.move(model_folder, model_folder.replace('_trained', '_evaluated'))
+
+        # Check for dnn.png in all 3 experiments.
+        if os.path.isfile(os.path.join(model_folder, 'scan_batteries', 'target_anechoic_cyst_5mm', 'target_5*', 'dnn.png')) \
+            and os.path.isfile(os.path.join(model_folder, 'scan_batteries', 'target_phantom_anechoic_cyst_2p5mm', 'target_5*', 'dnn.png')) \
+            and os.path.isfile(os.path.join(model_folder, 'scan_batteries', 'target_in_vivo', 'target_19', 'dnn.png')):
+
+            shutil.move(model_folder, model_folder.replace('_trained', '_evaluated'))
+        else:
+            raise "evaluate_models.py: dnn.png check failed."
